+++ app/controllers/players_controller.rb
@@ -1,0 +1,7 @@
+require_relative "base_controller"
+
+class PlayersController < BaseController
+
+  MODEL = Player
+
+end