+++ db/migrate/201408121534_rename_commitments_to_migrations.rb
@@ -1,0 +1,7 @@
+class RenameCommitmentsToMigrations < ActiveRecord::Migration
+
+  def change
+    rename_table :commitments, :memberships
+  end
+
+end