+++ db/migrate/201408121542_add_captain_id.rb
@@ -1,0 +1,9 @@
+class AddCaptainId < ActiveRecord::Migration
+
+  def change
+    change_table "players", force: true do |t|
+      t.references :team
+    end
+  end
+
+end