+++ db/migrate/201408121610_remove_stray_columns.rb
@@ -1,0 +1,9 @@
+class RemoveStrayColumns < ActiveRecord::Migration
+
+  def change
+    rename_column :players, :commitments_id, :memberships_id
+    rename_column :teams, :commitments_id, :memberships_id
+    remove_column :players, :team_id
+  end
+
+end