+++ spec/integration/deleting_sports_from_the_league_spec.rb
@@ -1,7 +1,9 @@
 RSpec.describe "Deleting sports from the league", :integration do
+  let!(:player){Player.create(username: "Greg", first_name: "Greg", last_name: "Jones", age: 37, is_male: 1)}
+  let!(:sport){Sport.create(name: "Baseball")}
+
   context "with valid input using sport index" do
-    let!(:sport){Sport.create(name: "Baseball")}
-    let!(:output){run_program_with_input("delete", "1", "y", "end")}
+    let!(:output){run_program_with_input("Greg", "delete", "1", "y", "end")}
 
     it "deletes the sport from the league's database" do
       expect(Sport.count).to eq(0)
@@ -13,8 +15,7 @@ RSpec.describe "Deleting sports from the league", :integration do
   end
 
   context "with valid input using sport name" do
-    let!(:sport){Sport.create(name: "Baseball")}
-    let!(:output){run_program_with_input("delete", "Baseball", "y", "end")}
+    let!(:output){run_program_with_input("Greg", "delete", "Baseball", "y", "end")}
 
     it "deletes the sport from the league's database" do
       expect(Sport.count).to eq(0)
@@ -26,9 +27,8 @@ RSpec.describe "Deleting sports from the league", :integration do
   end
 
   context "with invalid input" do
-    let!(:sport){Sport.create(name: "Baseball")}
-    let!(:output1){run_program_with_input("delete", "Basebal", "end")}
-    let!(:output2){run_program_with_input("delete", "2", "end")}
+    let!(:output1){run_program_with_input("Greg", "delete", "Basebal", "end")}
+    let!(:output2){run_program_with_input("Greg", "delete", "2", "end")}
 
     it "does not delete the sport from the league's database" do
       expect(Sport.count).to eq(1)